Make the forecast loaders also accept datetime objects (#33)
Currently the solcast detection doesn't detect solcast properly as while
in the serialised test code it was iso date strings, in the actual home
assistant instance it's a datetime object.

That meant that the tests all passed while the actual real running HA
instance couldn't find solcast forecasts.

This fixes it so that all the forecast parsers just take the outputs
that could be datetime objects and accept either an ISO string or a
datetime object.