Upgrade the configuration selectors to show sensors which are compatible with haeo (#34)
Currently the selectors are showing nothing for price sensors since they
are looking for device class monetary which isn't on the price sensors.
Additionally, some of the "power" forecast sensors are associated with
"energy" entities (e.g., solar forecasts often have power as an
attribute, but the sensor itself measures "total energy for the day").

This PR will go through all entities and find the ones that haeo would
be happy with (ones which have the correct units on either their state
or forecast) and provide those as options for the user.