Break up the element class and put the logic in the individual elements (#51)
The element class was trying to do everything and instead of
consolidating the code and making it easier, it just made it a whole lot
harder to understand what was going on.

This refactors so that each individual element is responsible for its
own math. All the base class does now is handle connection power.