Fix the validator to get the real data when validating (#114)
The validator was passing in an empty forecast times list to get the
data to validate the structure.
The battery used some of those values to initialise the state of charge.
The code then of course had an index error.

This PR updates the validator to get the full data when it tries to
validate the structure so that it reflects the actual elements that are
created.

Should fix #109

I also got carried away when I noticed some typing/string constant
issues