Swap to using explicit ordering connections to make better battery decisions (#142)
This pull request introduces major enhancements to the battery modelling
logic, enabling support for time-varying battery section capacities and
dynamic energy redistribution between battery sections. The changes
include the introduction of a new `BatteryBalanceConnection` model
element, significant updates to how battery sections and their
capacities are handled, and adjustments to cost/incentive calculations
for more realistic modelling.

The main reason for this upgrade is that we can vary the percentages on
each section. That will allow us to have things like a "minimum reserve
power" in the battery that changes based on time.

As a side effect it gets rid of the "cost based" ordering which is more
of a guideline to the optimiser and replaces it with actual constraints
which will enforce the ordering.