Simplify the schema system from reflection to individual files (#166)
I think I was trying to be too clever and generalise my flows too much.
All that it ended up doing was converting every time I wanted to do a
little thing in one place into a thing I need to do in every file.

This sets a consistent pattern for the element flows and then makes a
unique flow for each file. If common patterns emerge we can pull out
utility, but I want to change it to be a "uses" not "is" pattern.